Bringing this fork up-to-date with Deadrik/master

+++ src/API/com/bioxx/libnoise/module/source/Billow.java
@@ -16,7 +16,7 @@
 
  This is a port of libnoise ( http://libnoise.sourceforge.net/index.html ).  Original implementation by Jason Bevins
 
-*/
+ */
 
 package com.bioxx.libnoise.module.source;
 
@@ -45,7 +45,7 @@ public class Billow extends Module {
 	protected double lacunarity = DEFAULT_BILLOW_LACUNARITY;
 	protected NoiseQuality quality = DEFAULT_BILLOW_QUALITY;
 	protected double persistence = DEFAULT_BILLOW_PERSISTENCE;
-	protected int seed = DEFAULT_BILLOW_SEED;
+	protected long seed = DEFAULT_BILLOW_SEED;
 	protected int octaveCount = DEFAULT_BILLOW_OCTAVE_COUNT;
 
 	public Billow() {
@@ -95,11 +95,11 @@ public class Billow extends Module {
 		this.persistence = persistance;
 	}
 
-	public int getSeed() {
+	public long getSeed() {
 		return seed;
 	}
 
-	public void setSeed(int seed) {
+	public void setSeed(long seed) {
 		this.seed = seed;
 	}
 
@@ -108,45 +108,45 @@ public class Billow extends Module {
 		return 0;
 	}
 
-    @Override
-    public double GetValue(double x, double y, double z) {
-        double z1 = z;
-        double y1 = y;
-        double x1 = x;
-        double value = 0.0;
-        double signal;
-        double curPersistence = 1.0;
-        double nx, ny, nz;
-        int seed;
+	@Override
+	public double GetValue(double x, double y, double z) {
+		double z1 = z;
+		double y1 = y;
+		double x1 = x;
+		double value = 0.0;
+		double signal;
+		double curPersistence = 1.0;
+		double nx, ny, nz;
+		long seed;
 
-        x1 *= frequency;
-        y1 *= frequency;
-        z1 *= frequency;
+		x1 *= frequency;
+		y1 *= frequency;
+		z1 *= frequency;
 
-        for (int curOctave = 0; curOctave < octaveCount; curOctave++) {
+		for (int curOctave = 0; curOctave < octaveCount; curOctave++) {
 
-            // Make sure that these floating-point values have the same range as a 32-
-            // bit integer so that we can pass them to the coherent-noise functions.
-            nx = Utils.MakeInt32Range(x1);
-            ny = Utils.MakeInt32Range(y1);
-            nz = Utils.MakeInt32Range(z1);
+			// Make sure that these floating-point values have the same range as a 32-
+			// bit integer so that we can pass them to the coherent-noise functions.
+			nx = Utils.MakeInt32Range(x1);
+			ny = Utils.MakeInt32Range(y1);
+			nz = Utils.MakeInt32Range(z1);
 
-            // Get the coherent-noise value from the input value and add it to the
-            // final result.
-            seed = (this.seed + curOctave);
-            signal = Noise.GradientCoherentNoise3D(nx, ny, nz, seed, quality);
-            signal = 2.0 * Math.abs(signal) - 1.0;
-            value += signal * curPersistence;
+			// Get the coherent-noise value from the input value and add it to the
+			// final result.
+			seed = (this.seed + curOctave);
+			signal = Noise.GradientCoherentNoise3D(nx, ny, nz, seed, quality);
+			signal = 2.0 * Math.abs(signal) - 1.0;
+			value += signal * curPersistence;
 
-            // Prepare the next octave.
-            x1 *= lacunarity;
-            y1 *= lacunarity;
-            z1 *= lacunarity;
-            curPersistence *= persistence;
-        }
-        value += 0.5;
+			// Prepare the next octave.
+			x1 *= lacunarity;
+			y1 *= lacunarity;
+			z1 *= lacunarity;
+			curPersistence *= persistence;
+		}
+		value += 0.5;
 
-        return value;
-    }
+		return value;
+	}
 
 }
